requirements.txt
asgiref==3.5.2
contourpy==1.0.6
cycler==0.11.0
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.8.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.5
packaging==21.3
Pillow==9.3.0
platformdirs==2.5.4
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
virtualenv==20.17.0
